+++ Robust.Shared/Physics/Dynamics/ContactManager.cs
@@ -63,7 +63,6 @@ namespace Robust.Shared.Physics.Dynamics
                                                                Contact.ContactType.EdgeAndCircle,
                                                                Contact.ContactType.PolygonAndCircle,
                                                                Contact.ContactType.ChainAndCircle,
-                                                               Contact.ContactType.AabbAndCircle,
                                                            },
                                                            {
                                                                // Edge register
@@ -71,7 +70,6 @@ namespace Robust.Shared.Physics.Dynamics
                                                                Contact.ContactType.NotSupported, // Edge
                                                                Contact.ContactType.EdgeAndPolygon,
                                                                Contact.ContactType.NotSupported, // Chain
-                                                               Contact.ContactType.NotSupported, // Aabb
                                                            },
                                                            {
                                                                // Polygon register
@@ -79,7 +77,6 @@ namespace Robust.Shared.Physics.Dynamics
                                                                Contact.ContactType.EdgeAndPolygon,
                                                                Contact.ContactType.Polygon,
                                                                Contact.ContactType.ChainAndPolygon,
-                                                               Contact.ContactType.AabbAndPolygon,
                                                            },
                                                            {
                                                                // Chain register
@@ -87,15 +84,6 @@ namespace Robust.Shared.Physics.Dynamics
                                                                Contact.ContactType.NotSupported, // Edge
                                                                Contact.ContactType.ChainAndPolygon,
                                                                Contact.ContactType.NotSupported, // Chain
-                                                               Contact.ContactType.NotSupported, // Aabb - TODO Just cast to poly
-                                                           },
-                                                           {
-                                                               // Aabb register
-                                                               Contact.ContactType.AabbAndCircle,
-                                                               Contact.ContactType.NotSupported, // Edge - TODO Just cast to poly
-                                                               Contact.ContactType.AabbAndPolygon,
-                                                               Contact.ContactType.NotSupported, // Chain - TODO Just cast to poly
-                                                               Contact.ContactType.Aabb,
                                                            }
                                                        };
